modularizacion del main
